+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+In Development
+--------------
+- Add `access-denied` item to `--ok-errors`
+
 v0.1.0 (2025-02-04)
 -------------------
 - Add `--compress-filter-msgs` option